refactor(P2.6 slice 1): move #529 property-guard pair into render_plan/with_to_cte
First sub-slice of P2.6 (REFACTORING_SAFETY_PLAN.md §5.1) — begin moving the
WITH→CTE "entangled core" out of the ~13.3K-line plan_builder_utils.rs into a
new `render_plan/with_to_cte/` directory module, one byte-identical MOVE per PR.

Moved verbatim (old lines 10792–11185): the #529 shape-1 loud-guard property
helpers `table_role_dependent_property_names` + `collect_property_accesses` and
their `collect_property_accesses_tests` unit module. Landed in a single
`with_to_cte/mod.rs` (a direct child of `render_plan`, same module depth as
plan_builder_utils) so the test module's `super::super::render_expr::…`
path-expressions resolve byte-identically — a sub-file split would push `super`
one level deeper and force non-verbatim path edits (that is Phase-4
decomposition, out of scope for a move).

Both fns widened `fn` → `pub(crate) fn` at the move site (the sole non-verbatim
change) so the transition re-export in plan_builder_utils reaches them;
`build_chained_with_match_cte_plan`'s two call sites keep resolving unqualified
via `pub(crate) use super::with_to_cte::{…}`. The new module uses targeted
imports (P2.5 cte_rewrite precedent), with the test-only render types gated
behind `#[cfg(test)]` so the lib build stays warning-free.

Gates: 1612 lib + 529 integration (incl corpus_sweep + sql_golden) + 1 ratchet
+ 7 unit, 0 failures; goldens + corpus byte-identical (no churn); ratchet
net-zero; fmt/clippy clean. Next: P2.6 slice 2 (replace_v2).
